shader framework

+++ include/shared/linmath.h
@@ -4,7 +4,7 @@
 #include <math.h>
 
 #define LINMATH_H_DEFINE_VEC(n) \
-typedef float vec##n[n]; \
+typedef double vec##n[n]; \
 static inline void vec##n##_reset(vec##n r) \
 { \
 	int i; \
@@ -23,27 +23,27 @@ static inline void vec##n##_sub(vec##n r, vec##n const a, vec##n const b) \
 	for(i=0; i<n; ++i) \
 		r[i] = a[i] - b[i]; \
 } \
-static inline void vec##n##_scale(vec##n r, vec##n const v, float const s) \
+static inline void vec##n##_scale(vec##n r, vec##n const v, double const s) \
 { \
 	int i; \
 	for(i=0; i<n; ++i) \
 		r[i] = v[i] * s; \
 } \
-static inline float vec##n##_mul_inner(vec##n const a, vec##n const b) \
+static inline double vec##n##_mul_inner(vec##n const a, vec##n const b) \
 { \
-	float p = 0.; \
+	double p = 0.; \
 	int i; \
 	for(i=0; i<n; ++i) \
 		p += b[i]*a[i]; \
 	return p; \
 } \
-static inline float vec##n##_len(vec##n const v) \
+static inline double vec##n##_len(vec##n const v) \
 { \
 	return sqrtf(vec##n##_mul_inner(v,v)); \
 } \
 static inline void vec##n##_norm(vec##n r, vec##n const v) \
 { \
-	float k = 1.0 / vec##n##_len(v); \
+	double k = 1.0 / vec##n##_len(v); \
 	vec##n##_scale(r, v, k); \
 } \
 static inline void vec##n##_min(vec##n r, vec##n a, vec##n b) \
@@ -72,7 +72,7 @@ static inline void vec3_mul_cross(vec3 r, vec3 const a, vec3 const b)
 
 static inline void vec3_reflect(vec3 r, vec3 const v, vec3 const n)
 {
-	float p  = 2.f*vec3_mul_inner(v, n);
+	double p  = 2.f*vec3_mul_inner(v, n);
 	int i;
 	for(i=0;i<3;++i)
 		r[i] = v[i] - p*n[i];
@@ -88,7 +88,7 @@ static inline void vec4_mul_cross(vec4 r, vec4 a, vec4 b)
 
 static inline void vec4_reflect(vec4 r, vec4 v, vec4 n)
 {
-	float p  = 2.f*vec4_mul_inner(v, n);
+	double p  = 2.f*vec4_mul_inner(v, n);
 	int i;
 	for(i=0;i<4;++i)
 		r[i] = v[i] - p*n[i];
@@ -140,13 +140,13 @@ static inline void mat4x4_sub(mat4x4 M, mat4x4 a, mat4x4 b)
 	for(i=0; i<4; ++i)
 		vec4_sub(M[i], a[i], b[i]);
 }
-static inline void mat4x4_scale(mat4x4 M, mat4x4 a, float k)
+static inline void mat4x4_scale(mat4x4 M, mat4x4 a, double k)
 {
 	int i;
 	for(i=0; i<4; ++i)
 		vec4_scale(M[i], a[i], k);
 }
-static inline void mat4x4_scale_aniso(mat4x4 M, mat4x4 a, float x, float y, float z)
+static inline void mat4x4_scale_aniso(mat4x4 M, mat4x4 a, double x, double y, double z)
 {
 	int i;
 	vec4_scale(M[0], a[0], x);
@@ -176,14 +176,14 @@ static inline void mat4x4_mul_vec4(vec4 r, mat4x4 M, vec4 v)
 			r[j] += M[i][j] * v[i];
 	}
 }
-static inline void mat4x4_translate(mat4x4 T, float x, float y, float z)
+static inline void mat4x4_translate(mat4x4 T, double x, double y, double z)
 {
 	mat4x4_identity(T);
 	T[3][0] = x;
 	T[3][1] = y;
 	T[3][2] = z;
 }
-static inline void mat4x4_translate_in_place(mat4x4 M, float x, float y, float z)
+static inline void mat4x4_translate_in_place(mat4x4 M, double x, double y, double z)
 {
 	vec4 t = {x, y, z, 0};
 	vec4 r;
@@ -199,10 +199,10 @@ static inline void mat4x4_from_vec3_mul_outer(mat4x4 M, vec3 a, vec3 b)
 	for(i=0; i<4; ++i) for(j=0; j<4; ++j)
 		M[i][j] = i<3 && j<3 ? a[i] * b[j] : 0.f;
 }
-static inline void mat4x4_rotate(mat4x4 R, mat4x4 M, float x, float y, float z, float angle)
+static inline void mat4x4_rotate(mat4x4 R, mat4x4 M, double x, double y, double z, double angle)
 {
-	float s = sinf(angle);
-	float c = cosf(angle);
+	double s = sinf(angle);
+	double c = cosf(angle);
 	vec3 u = {x, y, z};
 
 	if(vec3_len(u) > 1e-4) {
@@ -233,10 +233,10 @@ static inline void mat4x4_rotate(mat4x4 R, mat4x4 M, float x, float y, float z, 
 		mat4x4_dup(R, M);
 	}
 }
-static inline void mat4x4_rotate_X(mat4x4 Q, mat4x4 M, float angle)
+static inline void mat4x4_rotate_X(mat4x4 Q, mat4x4 M, double angle)
 {
-	float s = sinf(angle);
-	float c = cosf(angle);
+	double s = sinf(angle);
+	double c = cosf(angle);
 	mat4x4 R = {
 		{1.f, 0.f, 0.f, 0.f},
 		{0.f,   c,   s, 0.f},
@@ -245,10 +245,10 @@ static inline void mat4x4_rotate_X(mat4x4 Q, mat4x4 M, float angle)
 	};
 	mat4x4_mul(Q, M, R);
 }
-static inline void mat4x4_rotate_Y(mat4x4 Q, mat4x4 M, float angle)
+static inline void mat4x4_rotate_Y(mat4x4 Q, mat4x4 M, double angle)
 {
-	float s = sinf(angle);
-	float c = cosf(angle);
+	double s = sinf(angle);
+	double c = cosf(angle);
 	mat4x4 R = {
 		{   c, 0.f,   s, 0.f},
 		{ 0.f, 1.f, 0.f, 0.f},
@@ -257,10 +257,10 @@ static inline void mat4x4_rotate_Y(mat4x4 Q, mat4x4 M, float angle)
 	};
 	mat4x4_mul(Q, M, R);
 }
-static inline void mat4x4_rotate_Z(mat4x4 Q, mat4x4 M, float angle)
+static inline void mat4x4_rotate_Z(mat4x4 Q, mat4x4 M, double angle)
 {
-	float s = sinf(angle);
-	float c = cosf(angle);
+	double s = sinf(angle);
+	double c = cosf(angle);
 	mat4x4 R = {
 		{   c,   s, 0.f, 0.f},
 		{  -s,   c, 0.f, 0.f},
@@ -271,8 +271,8 @@ static inline void mat4x4_rotate_Z(mat4x4 Q, mat4x4 M, float angle)
 }
 static inline void mat4x4_invert(mat4x4 T, mat4x4 M)
 {
-	float s[6];
-	float c[6];
+	double s[6];
+	double c[6];
 	s[0] = M[0][0]*M[1][1] - M[1][0]*M[0][1];
 	s[1] = M[0][0]*M[1][2] - M[1][0]*M[0][2];
 	s[2] = M[0][0]*M[1][3] - M[1][0]*M[0][3];
@@ -288,7 +288,7 @@ static inline void mat4x4_invert(mat4x4 T, mat4x4 M)
 	c[5] = M[2][2]*M[3][3] - M[3][2]*M[2][3];
 	
 	/* Assumes it is invertible */
-	float idet = 1.0f/( s[0]*c[5]-s[1]*c[4]+s[2]*c[3]+s[3]*c[2]-s[4]*c[1]+s[5]*c[0] );
+	double idet = 1.0f/( s[0]*c[5]-s[1]*c[4]+s[2]*c[3]+s[3]*c[2]-s[4]*c[1]+s[5]*c[0] );
 	
 	T[0][0] = ( M[1][1] * c[5] - M[1][2] * c[4] + M[1][3] * c[3]) * idet;
 	T[0][1] = (-M[0][1] * c[5] + M[0][2] * c[4] - M[0][3] * c[3]) * idet;
@@ -313,7 +313,7 @@ static inline void mat4x4_invert(mat4x4 T, mat4x4 M)
 static inline void mat4x4_orthonormalize(mat4x4 R, mat4x4 M)
 {
 	mat4x4_dup(R, M);
-	float s = 1.;
+	double s = 1.;
 	vec3 h;
 
 	vec3_norm(R[2], R[2]);
@@ -334,7 +334,7 @@ static inline void mat4x4_orthonormalize(mat4x4 R, mat4x4 M)
 	vec3_norm(R[0], R[0]);
 }
 
-static inline void mat4x4_frustum(mat4x4 M, float l, float r, float b, float t, float n, float f)
+static inline void mat4x4_frustum(mat4x4 M, double l, double r, double b, double t, double n, double f)
 {
 	M[0][0] = 2.f*n/(r-l);
 	M[0][1] = M[0][2] = M[0][3] = 0.f;
@@ -350,7 +350,7 @@ static inline void mat4x4_frustum(mat4x4 M, float l, float r, float b, float t, 
 	M[3][2] = -2.f*(f*n)/(f-n);
 	M[3][0] = M[3][1] = M[3][3] = 0.f;
 }
-static inline void mat4x4_ortho(mat4x4 M, float l, float r, float b, float t, float n, float f)
+static inline void mat4x4_ortho(mat4x4 M, double l, double r, double b, double t, double n, double f)
 {
 	M[0][0] = 2.f/(r-l);
 	M[0][1] = M[0][2] = M[0][3] = 0.f;
@@ -366,11 +366,11 @@ static inline void mat4x4_ortho(mat4x4 M, float l, float r, float b, float t, fl
 	M[3][2] = -(f+n)/(f-n);
 	M[3][3] = 1.f;
 }
-static inline void mat4x4_perspective(mat4x4 m, float y_fov, float aspect, float n, float f)
+static inline void mat4x4_perspective(mat4x4 m, double y_fov, double aspect, double n, double f)
 {
 	/* NOTE: Degrees are an unhandy unit to work with.
 	 * linmath.h uses radians for everything! */
-	float const a = 1.f / tan(y_fov / 2.f);
+	double const a = 1.f / tan(y_fov / 2.f);
 
 	m[0][0] = a / aspect;
 	m[0][1] = 0.f;
@@ -434,7 +434,7 @@ static inline void mat4x4_look_at(mat4x4 m, vec3 eye, vec3 center, vec3 up)
 	mat4x4_translate_in_place(m, -eye[0], -eye[1], -eye[2]);
 }
 
-typedef float quat[4];
+typedef double quat[4];
 static inline void quat_identity(quat q)
 {
 	q[0] = q[1] = q[2] = 0.f;
@@ -462,15 +462,15 @@ static inline void quat_mul(quat r, quat p, quat q)
 	vec3_add(r, r, w);
 	r[3] = p[3]*q[3] - vec3_mul_inner(p, q);
 }
-static inline void quat_scale(quat r, quat v, float s)
+static inline void quat_scale(quat r, quat v, double s)
 {
 	int i;
 	for(i=0; i<4; ++i)
 		r[i] = v[i] * s;
 }
-static inline float quat_inner_product(quat a, quat b)
+static inline double quat_inner_product(quat a, quat b)
 {
-	float p = 0.f;
+	double p = 0.f;
 	int i;
 	for(i=0; i<4; ++i)
 		p += b[i]*a[i];
@@ -483,7 +483,7 @@ static inline void quat_conj(quat r, quat q)
 		r[i] = -q[i];
 	r[3] = q[3];
 }
-static inline void quat_rotate(quat r, float angle, vec3 axis) {
+static inline void quat_rotate(quat r, double angle, vec3 axis) {
 	vec3 v;
 	vec3_scale(v, axis, sinf(angle / 2));
 	int i;
@@ -514,14 +514,14 @@ v' = v + q.w * t + cross(q.xyz, t)
 }
 static inline void mat4x4_from_quat(mat4x4 M, quat q)
 {
-	float a = q[3];
-	float b = q[0];
-	float c = q[1];
-	float d = q[2];
-	float a2 = a*a;
-	float b2 = b*b;
-	float c2 = c*c;
-	float d2 = d*d;
+	double a = q[3];
+	double b = q[0];
+	double c = q[1];
+	double d = q[2];
+	double a2 = a*a;
+	double b2 = b*b;
+	double c2 = c*c;
+	double d2 = d*d;
 	
 	M[0][0] = a2 + b2 - c2 - d2;
 	M[0][1] = 2.f*(b*c + a*d);
@@ -555,14 +555,14 @@ static inline void mat4x4o_mul_quat(mat4x4 R, mat4x4 M, quat q)
 }
 static inline void quat_from_mat4x4(quat q, mat4x4 M)
 {
-	float r=0.f;
+	double r=0.f;
 	int i;
 
 	int perm[] = { 0, 1, 2, 0, 1 };
 	int *p = perm;
 
 	for(i = 0; i<3; i++) {
-		float m = M[i][i];
+		double m = M[i][i];
 		if( m < r )
 			continue;
 		m = r;
